Add offset param to zlib.inflate.
Previously, to read each object from the packfile, we would call rest(),
which copies the rest of the packfile into a new array. Now, we just
give inflate the raw string and it will stop reading when it is
finished.

For now, the WebWorker inflating has been disabled. This change
drastically improves the performance of listing the log. However, I
haven't tested it on larger files yet, so it may still be useful to
delegate to a worker for reading objects.