Fix little bugs that were breaking the demo.
Fix HttpRemote.prototype.getObjectFromCachedPacks, which was failing to
call the callback when object wasn't in the cache.

Fix incorrect usage of this within a lambda in
TreeDiff.prototype.compareBlobEntry,

Fix FileDiff.prototype.toHtml, which was still calling Git.escapeHTML
instead of utils.escapeHTML.

Propagate errors up the callback chain in a few places.